Hypothesis Testing Paired Data
Have developed in .Rmd a fictitious dataset and graphical representation of t-tests for paired data, with all three posibilities of "alternative", i.e. "greater", "less", and "two.sided" with cases of rejection and acceptance of null hypothesis.  As a learner, I found it easier to sketch out curves, quantiles and the T-statistic to understand concepts of acceptance and rejection etc.  Hence, developed this example to help others too.